prov/verbs: Merge MSG EP and DGRAM EP routines into common one
