+++ prov/verbs/Makefile.include
@@ -24,7 +24,6 @@ _verbs_files =							\
 	prov/verbs/src/ep_rdm/verbs_utils.c			\
 	prov/verbs/src/ep_rdm/verbs_utils.h			\
 	prov/verbs/src/verbs_dgram.h				\
-	prov/verbs/src/verbs_dgram_ep.c				\
 	prov/verbs/src/verbs_dgram_ep_msg.c			\
 	prov/verbs/src/verbs_dgram_cq.c				\
 	prov/verbs/src/verbs_dgram_cntr.c			\
