+++ prov/verbs/src/verbs_domain.c
@@ -389,7 +389,7 @@ static struct fi_ops_domain fi_ibv_dgram_domain_ops = {
 	.size = sizeof(struct fi_ops_domain),
 	.av_open = fi_ibv_dgram_av_open,
 	.cq_open = fi_ibv_dgram_cq_open,
-	.endpoint = fi_ibv_dgram_endpoint_open,
+	.endpoint = fi_ibv_open_ep,
 	.scalable_ep = fi_no_scalable_ep,
 	.cntr_open = fi_ibv_dgram_cntr_open,
 	.poll_open = fi_no_poll_open,
